requirements.txt
alipay-sdk-python==3.3.398
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
Django==3.1.2
django-cors-headers==3.8.0
djangorestframework==3.12.4
drf-yasg2==1.19.4
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.1
packaging==21.0
Pillow==8.3.1
pyasn1==0.4.8
pycryptodome==3.11.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytz==2020.1
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
shellescape==3.8.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.26.6
